The soundlib_scheduler function now has uint8_t argument, aligned with the scheduler_usertask function.
- The soundlib_scheduler function now has uint8_t argument, aligned with the scheduler_usertask function.
- Added more text and comments.

+++ tinyavrlib/soundlib.h
@@ -30,7 +30,7 @@ typedef const uint16_t soundlib_melody_data_t;
 // ----------------------------------------------------------------------------
 
 void soundlib_init(void);
-void soundlib_scheduler(int);
+void soundlib_scheduler(uint8_t);
 void soundlib_tone_play(uint16_t, uint8_t);
 void soundlib_tone_stop(void);
 void soundlib_melody_play(const uint16_t[], int);
